Delegate to a object which does authentication
